requirements.txt
asgiref==3.3.1
certifi==2020.6.20
cloudinary==1.23.0
cycler==0.10.0
decorator==4.4.2
Django==3.1.3
gunicorn==20.0.4
imageio==2.9.0
imutils==0.5.3
kiwisolver==1.2.0
llvmlite==0.34.0
matplotlib==3.3.2
networkx==2.5
numba==0.51.2
numpy==1.19.2
opencv-contrib-python==4.2.0.32
Pillow==7.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
PyWavelets==1.1.1
scikit-image==0.17.2
scipy==1.5.2
six==1.15.0
sqlparse==0.4.1
tifffile==2020.10.1
urllib3==1.26.2
